+++ src/styles/index.css
@@ -44,4 +44,3 @@ body, html, #root {
   font-size: 1.2em;
   z-index: 2;
 }
-
