+++ src/types/person.ts
@@ -1,0 +1,9 @@
+export interface Person {
+    name: string;
+    lastName: string;
+    address: string;
+    city: string;
+    tel: string;
+    url: string;
+    email: string;
+}
